Refactor SearchBar component to use let instead of const for search input value
The SearchBar component in SearchBar.tsx has been refactored to use the let keyword instead of const for the search input value. This change allows the value to be updated dynamically as the user types in the input field. Previously, the value was declared as a const, which prevented it from being reassigned. This commit ensures that the search input value is properly updated and stored in the state variable.